main Activity main page UI design
I have commented out sensorButton as I am not sure what it does. I don't know if it should go with the other buttons or in a corner on its own for example. I'll put it back once you tell me exactly what to do with it.

"Take me back" is also commented out. I will put it back later.