If the number of samples is less than 2 or 1 (don't remember what I set it to), it won't prompt the user if he wants to save his route. This is to prevent saving routes that have no coordinates.
Redid the route manager. Hiking should display the area in which the route is located, in case the user went to the countryside to hike there. Running and Biking should display the best time.
How to test?
Create a bunch of different routes, test in the route manager.
To do:
delete button color

Proximity check:
nearby running route: less than 1KM from current location
nearby hiking: less than 10KM from current location
nearby biking: less than 3KM
We could adjust these values for the demo, or not.

How to test?
There is no easy way to test but here is an indirect way:
on line 232 in routeManager.java
if(distance < DISTANCE_THRESHOLD)
change the < to a >
this will reverse the condition for a route to be considered nearby, so nearby routes are now considered far. compile the code, in the route manager click on nearby. you shouldn't see anything.